refactor: Layout-associated hooks relocated to layout.hooks
For the purpose of improved maintainability, all hooks related to the
Layout component have now been moved to layout.hooks. The intention of
this rearrangement is to gather all correlated files in one place,
thereby simplifying future changes. Additionally, necessary adjustments
have been applied to all subscribing entities to accommodate this shift.
