+++ components/layout/app/appNavigation/index.tsx
@@ -1,7 +1,6 @@
 import { DisableButton } from '@buttons/disableButton';
 import { IconButton } from '@buttons/iconButton';
 import { ICON_ADD_TASK } from '@data/materialSymbols';
-import { useNavigationOpen } from '@hooks/layouts';
 import { useConditionCheckCreateModalOpen } from '@hooks/misc';
 import { useTodoModalStateOpen } from '@hooks/modals';
 import { Logo } from '@layout/layoutHeader/logo';
@@ -15,6 +14,7 @@ import { AppSidebarMenu } from './appSidebarMenu';
 import dynamic from 'next/dynamic';
 import { LoadingLabels } from '@components/loadable/loadingStates/loadingLabels';
 import { SvgIcon } from '@icon/svgIcon';
+import { useNavigationOpen } from '@layout/layout.hooks';
 
 const LabelList = dynamic(() => import('@label/labelList').then((mod) => mod.LabelList), {
   ssr: false,
