+++ components/layout/layoutHeader/navigationButton.tsx
@@ -1,5 +1,5 @@
 import { IconButton } from '@buttons/iconButton';
-import { useNavigationOpen } from '@hooks/layouts';
+import { useNavigationOpen } from '@layout/layout.hooks';
 import { optionsButtonSidebarToggle } from '@options/button';
 
 export const NavigationButton = () => {
